Update shop server
1. 프론트 회의 결과 tag -> thema 로 통일

2. thema는 리스트 뽑을 때만 사용, detail view에는 쏴주지 않음

3. Shop List Read 쿼리 수정

4. 조건문 오류 수정

+++ Backend/Shop/models/info.js
@@ -32,18 +32,18 @@ module.exports = (sequelize, DataTypes) => {
         },
     
         grade_avg: {
-            type: DataTypes.DOUBLE,
+            type: DataTypes.FLOAT,
             allowNULL: true,
             defaultValue: 0
         },
 
         latitude: {
-            type: DataTypes.DOUBLE,
+            type: DataTypes.FLOAT,
             allowNULL: true
         },
         
         longitude: {
-            type: DataTypes.DOUBLE,
+            type: DataTypes.FLOAT,
             allowNULL: true
         },
